bugfix: geometry module, all rotation-matrix related angles had wrong sign
The wrong sign was based on a stupid bug in the elementary rotation
matrices, which led to transposed arrays (on the last two dimensions),
i.e., wrong sign for rotation matrices as R.T = R^-1. As all other
functions were tested against the element matrices, a lot of wrong signs
were added to the code.